refactor: DataSource와 Repository의 역할 재정의
- DataSource는 데이터를 가져오고 전달하는 역할의 객체로 둔다.
- Repository에서 필요로되는 데이터를 파싱하고 적절한 데이터를 전달한다.
- 이 역할을 다시 정의하며 코드를 수정한다.
