+++ app/src/main/java/com/woowa/banchan/data/local/datasource/recent/RecentDataSource.kt
@@ -5,5 +5,6 @@ import com.woowa.banchan.data.local.entity.RecentDto
 interface RecentDataSource {
 
     suspend fun getRecentList(): Result<List<RecentDto>>
+    suspend fun deleteRecent(recentDto: RecentDto): Result<Unit>
     suspend fun insertRecent(recentDto: RecentDto): Result<Unit>
 }
