+++ app/src/main/java/com/woowa/banchan/data/local/datasource/recent/RecentDataSourceImpl.kt
@@ -2,7 +2,6 @@ package com.woowa.banchan.data.local.datasource.recent
 
 import com.woowa.banchan.data.local.dao.RecentDao
 import com.woowa.banchan.data.local.entity.RecentDto
-import java.util.*
 import javax.inject.Inject
 
 class RecentDataSourceImpl @Inject constructor(
@@ -10,20 +9,10 @@ class RecentDataSourceImpl @Inject constructor(
 ) : RecentDataSource {
 
     override suspend fun getRecentList(): Result<List<RecentDto>> =
-        runCatching {
-            val list = recentDao.getRecentList().toMutableList()
-            val curCalendar = Calendar.getInstance()
-                .apply { time = Date() }
-            val calendar = Calendar.getInstance()
-            val retList = mutableListOf<RecentDto>()
-            list.forEach {
-                calendar.time = it.time
-                if (calendar.get(Calendar.DAY_OF_WEEK) != curCalendar.get(Calendar.DAY_OF_WEEK))
-                    recentDao.delete(it)
-                else retList.add(it)
-            }
-            retList
-        }
+        runCatching { recentDao.getRecentList() }
+
+    override suspend fun deleteRecent(recentDto: RecentDto): Result<Unit> =
+        runCatching { recentDao.delete(recentDto) }
 
     override suspend fun insertRecent(recentDto: RecentDto): Result<Unit> =
         runCatching { recentDao.insert(recentDto) }
